Separate into mock, live, and both
